Module Nocrypto

Nocrypto: for when you're sick of crypto.

A compact crypto library.

The overarching API principle is simply mapping inputs to outputs, wherever feasible.

Similar algorithms in the same class (like hashes or block ciphers) are presented as distinct modules sharing the same signature.

Randomness is treated as an ambient effect.

Utilities

module Base64 : sig ... end

Base64 conversion.

module Numeric : sig ... end

Numeric utilities.

Hashing

module Hash : sig ... end

Hashes.

Symmetric-key cryptography

module Cipher_block : sig ... end

Block ciphers.

module Cipher_stream : sig ... end

Streaming ciphers.

Randomness

module Rng : sig ... end

Secure random number generation.

Public-key cryptography

module Rsa : sig ... end

RSA public-key cryptography.

module Dsa : sig ... end

DSA digital signature algorithm.

module Dh : sig ... end

Diffie-Hellman, MODP version.