Module Nocrypto.Cipher_block

Block ciphers.

Each algorithm, and each mode of operation, is contained in its own separate module.

module S : sig ... end

Module types for various block cipher modes of operation.

module Counter : sig ... end

BE counter function.

module AES : sig ... end

module DES : sig ... end