Module DES.CBC

type key

val of_secret : Cstruct.t -> key

val key_sizes : int array

val block_size : int

val next_iv : iv:Cstruct.t -> Cstruct.t -> Cstruct.t

next_iv iv ciphertext for a ciphertext and an iv it was computed with is the iv to use to encrypt the next message, for protocols which perform inter-message chaining. It is either the last block of ciphertext or iv if msg is too short.

val encrypt : key:key -> iv:Cstruct.t -> Cstruct.t -> Cstruct.t

val decrypt : key:key -> iv:Cstruct.t -> Cstruct.t -> Cstruct.t