Module Nocrypto.Numeric

Numeric utilities.

module type S = sig ... end

Augmented numeric type. Includes basic common numeric ops, range of conversions to and from variously-sized int types, and a few basic function for representing such numbers as Cstruct.t.

module Int : S with type t = int

module Int32 : S with type t = int32

module Int64 : S with type t = int64

module Z : S with type t = Z.t

Misc elementary number theory

val pseudoprime : Z.t -> bool

Miller-Rabin with sane rounds parameter.