Module Rsa.OAEP

Parameters

module H : Hash.S

Signature

val encrypt : ?⁠g:Rng.g -> ?⁠label:Cstruct.t -> key:pub -> Cstruct.t -> Cstruct.t

encrypt ~g ~label ~key message is OAEP-padded and encrypted message, using the optional label.

raises Insufficient_key

(see Insufficient_key)

val decrypt : ?⁠mask:mask -> ?⁠label:Cstruct.t -> key:priv -> Cstruct.t -> Cstruct.t option

decrypt ~mask ~label ~key ciphertext is Some message if the ciphertext was produced by the corresponding encrypt operation, or None otherwise.